Do apply patches instantly
When using patch_state API, in order to prevent the server-side state to
diverge from client-side state on invalid states we should crash the
channel if prepared patch is invalid and cannot be applied.

We also now apply the patches instantly so that the server can see the
full patched state in on_message callbacks, after patches have been
applied but before they've been broadcasted.